requirements.txt
asgiref==3.11.1
dj-database-url==3.0.1
Django==4.2.29
gunicorn==23.0.0
packaging==26.0
psycopg==3.2.13
psycopg-binary==3.2.13
python-dotenv==1.2.1
sqlparse==0.5.5
typing_extensions==4.15.0
whitenoise==6.11.0
